feat: 해당 친구가 있는 journal 불러오기
